Remove NoMethodError test
This test causes an error in rspec >= 3 due to a deprecation of the `not_to
raise_error` syntax. The caller is instead expected to either specify that no
error will be raised, or give a specific error that will be raised.

Since the second case is covered by another test, this one can be removed.